호스트 공간 관리 페이지 수정(검색기능 및 코드수정), 호스트 상품 관리 페이지 수정, 상세보기, 삭제, 검색기능

+++ src/main/java/com/skillspace/sgs/host/space/HostSpaceController.java
@@ -19,7 +19,6 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import com.skillspace.sgs.admin.category.AdminCategoryService;
 import com.skillspace.sgs.common.utils.PageMaker;
 import com.skillspace.sgs.common.utils.SearchCriteria;
-import com.skillspace.sgs.common.utils.SearchItem;
 import com.skillspace.sgs.guest.GuestDTO;
 
 import jakarta.servlet.http.HttpSession;
@@ -35,25 +34,29 @@ public class HostSpaceController {
 	private final AdminCategoryService adminCategoryService;
 	private final HostSpaceService hostSpaceService;
 	
+	private String spaceListUrl = "/host/space/spaceList";
+	
 	// 호스트 공간 목록
 	@GetMapping("/spaceList")
 	public void spaceList(SearchCriteria cri, 
-			@ModelAttribute("searchItem") SearchItem searchItem,
 			Model model, 
 			HttpSession session) throws Exception {
 		
 		log.info("검색 정보 : " + cri);
-		log.info("검색아이템 : " + searchItem);
 		
 		// 1) 유저 공간 정보
 		String user_id = ((GuestDTO)session.getAttribute("login_auth")).getUser_id();
 		// 2) 공간 리스트 정보 가져오기(이미지 계층형 데이터로)
+		
+		if (cri.getPerPageNum() == 0) { // perPageNum이 0 인 경우 (초기 요청)
+	        cri.setPerPageNum(5);	// 페이지 관리 상수값
+	    }
 		PageMaker pageMaker = new PageMaker();
 		pageMaker.setDisplayPageNum(2);
 		pageMaker.setCri(cri);
-		pageMaker.setTotalCount(hostSpaceService.getCountSpaceByUser_id(user_id));
+		pageMaker.setTotalCount(hostSpaceService.getCountSpaceByUser_id(user_id, cri));
 		
-		List<HostSpaceDTO> spaceList = hostSpaceService.sapceList(user_id, cri, searchItem);
+		List<HostSpaceDTO> spaceList = hostSpaceService.sapceList(user_id, cri);
 		spaceList.forEach(space_info -> {
 			space_info.getImages().forEach(image_info -> {
 				image_info.setImage_up_folder(image_info.getImage_up_folder().replace("\\", "/"));
@@ -87,7 +90,7 @@ public class HostSpaceController {
 //		if(guestDTO.getRole().equals("H")) {
 //		}
 		
-		return "redirect:/host/space/spaceList";
+		return "redirect:" + spaceListUrl;
 	}
 	
 	// 호스트 공간 정보 수정 폼
@@ -113,36 +116,33 @@ public class HostSpaceController {
 			RedirectAttributes rttr) throws Exception {
 		
 		// 1) 로그인 호스트
-		GuestDTO guestDTO = ((GuestDTO)session.getAttribute("login_auth"));
-		dto.setUser_id(guestDTO.getUser_id());
+		if(!isProductOwner(dto.getHost_space_id(), session)) {
+			rttr.addFlashAttribute("errorMessage", "수정 권한이 없습니다.");
+    		return "redirect:" + spaceListUrl;
+		}
 		
 		// 2) 호스트 정보 수정
 		hostSpaceService.modifyHostSpace(dto, imageFiles, loadImages, changeImages);
 		
 		// 원래 상태의 목록으로 주소 이동 작업.
-		
-		rttr.addAttribute("page", cri.getPage());
-		rttr.addAttribute("perPageNum", cri.getPerPageNum());
-		rttr.addAttribute("searchType", cri.getSearchType());
-		rttr.addAttribute("keyword", cri.getKeyword());
-		return "redirect:/host/space/spaceList";
+		addSearchAndPagingParams(cri, rttr);
+		return "redirect:" + spaceListUrl;
 	}
 	
 	// 호스트 공간 삭제
 	@GetMapping("/deleteSpace")
-	public String deleteSpace(
-			Integer host_space_id, 
-			@ModelAttribute("cri") SearchCriteria cri,
-			RedirectAttributes rttr) throws Exception {
+	public String deleteSpace(Integer host_space_id, SearchCriteria cri, 
+			HttpSession session, RedirectAttributes rttr) throws Exception {
+		
+		if(!isProductOwner(host_space_id, session)) {
+			rttr.addFlashAttribute("errorMessage", "삭제 권한이 없습니다.");
+    		return "redirect:" + spaceListUrl;
+		}
 		
 		hostSpaceService.deleteHostSpace(host_space_id);
 		
-		
-		rttr.addAttribute("page", cri.getPage());
-		rttr.addAttribute("perPageNum", cri.getPerPageNum());
-		rttr.addAttribute("searchType", cri.getSearchType());
-		rttr.addAttribute("keyword", cri.getKeyword());
-		return "redirect:/host/space/spaceList";
+		addSearchAndPagingParams(cri, rttr);
+		return "redirect:" + spaceListUrl;
 	}
 	
 	// 체크박스 선택된 공간 삭제
@@ -154,6 +154,24 @@ public class HostSpaceController {
 		hostSpaceService.selectedDelete(host_space_id_arr);
 		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
+	
+	// 상품 유저 아이디와 로그인 유저 아이디 비교
+    private boolean isProductOwner(int host_space_id, HttpSession session) throws Exception {
+		GuestDTO guestDTO = ((GuestDTO)session.getAttribute("login_auth"));
+		String space_user_id = hostSpaceService.getHostSpaceUserIdById(host_space_id);
+		return guestDTO.getUser_id().equals(space_user_id);
+    }
+
+    // 검색 조건 및 페이징 정보 유지
+    private void addSearchAndPagingParams(SearchCriteria cri, RedirectAttributes rttr) {
+		rttr.addAttribute("page", cri.getPage());
+		rttr.addAttribute("perPageNum", cri.getPerPageNum());
+		rttr.addAttribute("searchType", cri.getSearchType());
+		rttr.addAttribute("keyword", cri.getKeyword());
+		rttr.addAttribute("start_date", cri.getStart_date());
+		rttr.addAttribute("end_date", cri.getEnd_date());
+		rttr.addAttribute("visible_status", cri.getVisible_status());
+    }
